// src/common/enums/tipo-erro.enum.ts
export enum TipoErro {
  SEM_ERROS = -1,
  USUARIO_SEM_PERMISSAO = 0,
  SEM_AUTENTICACAO = 1,
  DOCUMENTO_NAO_ENCONTRADO = 2,
  ERRO_CONEXAO_BD = 3,
  SERVICO_INDISPONIVEL = 4,
  ERRO_DE_COMUNICACAO = 5,
  ID_NAO_ENCONTRADO = 6,
  ERROR_AO_SALVAR = 7,
  ERROR_AO_ATUALIZAR = 8,
  ERROR_AO_DELETAR = 9,
  ERROR_DESCONHECIDO = 10,
  USUARIO_JA_EXISTE = 11,
  DADOS_INVALIDOS = 12,
}